Split roll forward test to three separate files
